+++ src/r-wrappers.h
@@ -20,7 +20,8 @@ SEXP SXP_get_optimal_possible_GEBV(SEXP exd, SEXP s_groups, SEXP s_eff_set);
 SEXP SXP_get_minimal_GEBV(SEXP exd, SEXP s_eff_set);
 
 SEXP SXP_find_crossovers(SEXP exd, SEXP s_parentFile, SEXP s_outFile, SEXP s_windowSize, SEXP s_certainty);
-SEXP SXP_send_map(SEXP exd, SEXP s_map);
+SEXP SXP_see_map(SEXP exd, SEXP s_map);
+SEXP SXP_see_effects(SEXP exd, SEXP s_effset, SEXP s_format);
 
 /*----------------------- Data Access ---------------------*/
 SEXP SXP_see_group_data(SEXP exd, SEXP s_group, SEXP s_whatData, SEXP s_eff_set_id, SEXP s_label_id);